+++ XMonad/Hooks/StatusBar/PP.hs
@@ -38,7 +38,8 @@ module XMonad.Hooks.StatusBar.PP (
     -- * Predicates and formatters
     -- $predicates
     WS(..), WSPP, WSPP', fallbackPrinters,
-    isUrgent, isCurrent, isVisible, isVisibleNoWindows, isHidden,
+    isUrgent, isCurrent, isCurrentNoWindows, isVisible, isVisibleNoWindows,
+    isHidden, hasWindows,
 
     -- * Example formatters
     dzenPP, xmobarPP, sjanssenPP, byorgeyPP,
@@ -96,6 +97,9 @@ data PP = PP { ppCurrent :: WorkspaceId -> String
                -- ^ how to print tags of empty hidden workspaces
              , ppVisibleNoWindows :: Maybe (WorkspaceId -> String)
                -- ^ how to print tags of empty visible workspaces
+             , ppCurrentNoWindows :: Maybe (WorkspaceId -> String)
+               -- ^ how to print the tag of the currently focused
+               -- workspace if the workspace is empty
              , ppUrgent :: WorkspaceId -> String
                -- ^ format to be applied to tags of urgent workspaces.
              , ppRename :: String -> WindowSpace -> String
@@ -156,6 +160,7 @@ data PP = PP { ppCurrent :: WorkspaceId -> String
 -- workspace layout, and the title of the focused window.
 instance Default PP where
   def = PP { ppCurrent          = wrap "[" "]"
+           , ppCurrentNoWindows = Nothing
            , ppVisible          = wrap "<" ">"
            , ppHidden           = id
            , ppHiddenNoWindows  = const ""
@@ -273,6 +278,7 @@ type WSPP = WSPP' (WorkspaceId -> String)
 fallbackPrinters :: WSPP
 fallbackPrinters = isUrgent            ?-> ppUrgent
                <|> isCurrent'          ?-> ppCurrent
+               <|> isCurrentNoWindows' ?-> liftA2 fromMaybe ppCurrent ppCurrentNoWindows
                <|> isVisible'          ?-> ppVisible
                <|> isVisibleNoWindows' ?-> liftA2 fromMaybe ppVisible ppVisibleNoWindows
                <|> isHidden'           ?-> ppHidden
@@ -287,16 +293,28 @@ isUrgent WS{..} = any (\x -> (== Just (S.tag wsWS)) (S.findTag x wsWindowSet)) w
 -- | Predicate for the current workspace. Caution: assumes default
 -- precedence is respected.
 isCurrent' :: WS -> Bool
-isCurrent' WS{..} = S.tag wsWS == S.currentTag wsWindowSet
+isCurrent' = isCurrentNoWindows' <&&> hasWindows
 
 -- | Predicate for the current workspace.
 isCurrent :: WS -> Bool
 isCurrent = (not <$> isUrgent) <&&> isCurrent'
 
+-- | Predicate for the current workspace with no windows. Caution:
+-- assumes default precedence is respected.
+isCurrentNoWindows' :: WS -> Bool
+isCurrentNoWindows' WS{..} = S.tag wsWS == S.currentTag wsWindowSet
+
+-- | Predicate for the current workspace with no windows.
+isCurrentNoWindows :: WS -> Bool
+isCurrentNoWindows =
+    (not <$> isUrgent)
+        <&&> (not <$> hasWindows)
+        <&&> isCurrentNoWindows'
+
 -- | Predicate for visible workspaces. Caution: assumes default
 -- precedence is respected.
 isVisible' :: WS -> Bool
-isVisible' = isVisibleNoWindows' <&&> isJust . S.stack . wsWS
+isVisible' = isVisibleNoWindows' <&&> hasWindows
 
 -- | Predicate for visible workspaces.
 isVisible :: WS -> Bool
@@ -312,14 +330,14 @@ isVisibleNoWindows' WS{..} = S.tag wsWS `elem` visibles
 isVisibleNoWindows :: WS -> Bool
 isVisibleNoWindows =
     (not <$> isUrgent)
-        <&&> (not <$> isCurrent')
-        <&&> (not <$> isVisible')
+        <&&> (not <$> isCurrentNoWindows')
+        <&&> (not <$> hasWindows)
         <&&> isVisibleNoWindows'
 
 -- | Predicate for non-empty hidden workspaces. Caution: assumes default
 -- precedence is respected.
 isHidden' :: WS -> Bool
-isHidden' = isJust . S.stack . wsWS
+isHidden' = hasWindows
 
 -- | Predicate for hidden workspaces.
 isHidden :: WS -> Bool
@@ -329,6 +347,10 @@ isHidden =
         <&&> (not <$> isVisible')
         <&&> (not <$> isVisibleNoWindows')
         <&&> isHidden'
+
+-- | Predicate for workspaces with windows.
+hasWindows :: WS -> Bool
+hasWindows = isJust . S.stack . wsWS
 
 pprWindowSetXinerama :: WindowSet -> String
 pprWindowSetXinerama ws = "[" ++ unwords onscreen ++ "] " ++ unwords offscreen